+++ commands/mlsd.js
@@ -32,6 +32,7 @@ command.add('MLSD', 'MLSD [<sp> pathname]', function (pathname, commandChannel, 
           var facts = '';
           facts += 'modify=' + facter.modify(stat);
           facts += ';perm=' + facter.perm(stat, session.user);
+          facts += ';size=' + stat.size;
 
           if (stat.isFile()) {
             facts += ';type=file';