+++ commands/list.js
@@ -1,9 +1,13 @@
 'use strict';
 
 // Local dependencies
-var command = require('../lib/command');
-var channel = require('../lib/data-channel');
-var fs      = require('../lib/fs');
+var facter      = require('../lib/facts');
+var command     = require('../lib/command');
+var dataChannel = require('../lib/data-channel');
+var fs          = require('../lib/fs');
+
+// Third party dependencies
+var path = require('path');
 
 /**
  * A LIST or NLST request asks the server to send the contents of a directory
@@ -92,61 +96,47 @@ var fs      = require('../lib/fs');
  * client.
  */
 command.add('LIST', 'LIST [<sp> pathname]', function (pathname, commandChannel, session) {
-  // Create a new data channel
-  var success = channel.create(session, function (socket, done) {
-    // Read in the files/directories in the user's current working
-    // directory
-    fs.readdir(fs.toAbsolute(pathname, session.cwd), function (err, files) {
-      if (files === null) {
-        files = [];
-      }
-
-      for (var i = 0; i < files.length; i++) {
-        var stat   = fs.statSync(session.cwd + '/' + files[i]);
-        var date   = stat.mtime;
-        var months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
-        var mode   = '';
-        var omode  = parseInt(stat.mode.toString(8), 10);
-        omode      = omode.toString().substr(omode.toString().length - 3);
-
-        if (stat.isFile()) {
-          mode += '-';
-        } else {
-          mode += 'd';
+  fs.readdir(fs.toAbsolute(pathname, session.cwd), function (err, files) {
+    if (err) {
+      commandChannel.write(550, fs.errorMessage(err, pathname));
+    } else {
+      var success = dataChannel.create(session, function (socket, done) {
+        if (files === null) {
+          files = [];
         }
 
-        for (var j = 0; j < 3; j++) {
-          if (omode[j] === '0') {
-            mode += '---';
-          } else if (omode[j] === '1') {
-            mode += '--x';
-          } else if (omode[j] === '2') {
-            mode += '-w-';
-          } else if (omode[j] === '3') {
-            mode += '-wx';
-          } else if (omode[j] === '4') {
-            mode += 'r--';
-          } else if (omode[j] === '5') {
-            mode += 'r-x';
-          } else if (omode[j] === '6') {
-            mode += 'rw-';
-          } else if (omode[j] === '7') {
-            mode += 'rwx';
+        for (var i = 0; i < files.length; i++) {
+          var stat     = fs.statSync(fs.toAbsolute(files[i], session.cwd));
+          var eplfLine = '+';
+
+          eplfLine += 'i' + facter.unique(stat);
+          eplfLine += ',m' + stat.mtime.valueOf();
+          eplfLine += ',s' + stat.size;
+          eplfLine += ',up' + stat.mode.toString(8).substring(stat.mode.toString(8) - 3);
+
+          if (stat.isFile()) {
+            eplfLine += ',r';
+          } else {
+            eplfLine += ',/';
           }
+
+          eplfLine += '\t';
+          eplfLine += path.basename(files[i]);
+          eplfLine += '\r\n';
+
+          socket.write(eplfLine);
         }
 
-        socket.write(mode + ' 1 ' + stat.uid + ' ' + stat.gid + '  ' + stat.size + ' ' + months[date.getMonth()] + ' ' + date.getDate() + ' ' + date.getFullYear() + ' ' + files[i] + '\n');
+        commandChannel.write(226, 'Directory sent OK.');
+
+        done();
+      });
+
+      if (!success) {
+        commandChannel.write(425, 'Unable to build data connection: Invalid argument');
+      } else {
+        commandChannel.write(150, 'Here comes the directory listing.');
       }
-
-      commandChannel.write(226, 'Directory sent OK.');
-
-      done();
-    });
+    }
   });
-
-  if (!success) {
-    commandChannel.write(425, 'Unable to build data connection: Invalid argument');
-  } else {
-    commandChannel.write(150, 'Here comes the directory listing.');
-  }
 });